save card design for action (like, note-writing)
